+++ src/main/java/net/spellcraftgaming/rpghud/gui/hud/element/hotbar/HudElementFoodHotbar.java
@@ -37,12 +37,12 @@ public class HudElementFoodHotbar extends HudElementBarred {
 			if (bonusHunger > foodMax)
 				bonusHunger = foodMax;
 			int colorPreview = offsetColor(this.settings.color_stamina, OFFSET_PREVIEW);
-			drawCustomBar(posX, height - 26, 200, 10, bonusHunger / foodMax * 100.0D, -1, -1, colorPreview, offsetColorPercent(colorPreview, OFFSET_PERCENT));
+			drawCustomBar(posX, height - 26, 200, 10, bonusHunger / (double)foodMax * 100.0D, -1, -1, colorPreview, offsetColorPercent(colorPreview, OFFSET_PERCENT));
 		}
 		if (GameData.isPlayerHungered()) {
-			drawCustomBar(posX, height - 26, 200, 10, stamina / foodMax * 100.0D, -1, -1, this.settings.color_hunger, offsetColorPercent(this.settings.color_hunger, OFFSET_PERCENT));
+			drawCustomBar(posX, height - 26, 200, 10, stamina / (double)foodMax * 100.0D, -1, -1, this.settings.color_hunger, offsetColorPercent(this.settings.color_hunger, OFFSET_PERCENT));
 		} else {
-			drawCustomBar(posX, height - 26, 200, 10, stamina / foodMax * 100.0D, -1, -1, this.settings.color_stamina, offsetColorPercent(this.settings.color_stamina, OFFSET_PERCENT));
+			drawCustomBar(posX, height - 26, 200, 10, stamina / (double)foodMax * 100.0D, -1, -1, this.settings.color_stamina, offsetColorPercent(this.settings.color_stamina, OFFSET_PERCENT));
 		}
 		String staminaString = stamina + "/" + foodMax;
 		if (this.settings.show_numbers_stamina)
